+++ WebApplication8/Controllers/usuariosController.cs
@@ -20,47 +20,12 @@ namespace WebApplication8.Controllers
 
         // GET: Usuarios
         // GET: Users
-		public Boolean isAdminUser()
-		{
-			if (User.Identity.IsAuthenticated)
-			{
-				var user = User.Identity;
-				ApplicationDbContext context = new ApplicationDbContext();
-				var UserManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(context));
-				var s = UserManager.GetRoles(user.GetUserId());
-				if (s[0].ToString() == "Admin")
-				{
-					return true;
-				}
-				else
-				{
-					return false;
-				}
-			}
-			return false;
-		}
+		
         public ActionResult Index()
         {
             var usuario = db.Usuarios.Include(u => u.Usuario_Moto);
             return View(usuario.ToList());
-            if (User.Identity.IsAuthenticated)
-            {
-                var user = User.Identity;
-                ViewBag.Name = user.Name;
-
-                ViewBag.displayMenu = "No";
-
-                if (isAdminUser())
-                {
-                    ViewBag.displayMenu = "Yes";
-                }
-                return View();
-            }
-            else
-            {
-                ViewBag.Name = "Not Logged IN";
-            }
-            return View();
+            
         }
 
         // GET: Usuarios/Details/5
@@ -154,9 +119,9 @@ namespace WebApplication8.Controllers
         // POST: Usuarios/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int id_Usuario)
         {
-            Usuario usuario = db.Usuarios.Find(id);
+            Usuario usuario = db.Usuarios.Find(id_Usuario);
             db.Usuarios.Remove(usuario);
             db.SaveChanges();
             return RedirectToAction("Index");
