+++ WebApplication8/Views/Account/ConfirmEmail.cshtml
@@ -5,6 +5,6 @@
 <h2>@ViewBag.Title.</h2>
 <div>
     <p>
-        Thank you for confirming your email. Please @Html.ActionLink("Click here to Log in", "Login", "Account", routeValues: null, htmlAttributes: new { id = "loginLink" })
+       Gracia por confirmar su Correo. Por Favo @Html.ActionLink("Click aqí para Inicicar Sesión", "Login", "Account", routeValues: null, htmlAttributes: new { id = "loginLink" })
     </p>
 </div>
