+++ WebApplication8/Views/Account/Login.cshtml
@@ -50,13 +50,7 @@
                         @Html.ActionLink("Forgot your password?", "ForgotPassword")
                     </p>
 
-
-
-
-
-                    }
-
-
+           }
 
 </section>
     </div>
