+++ WebApplication8/Views/Role/Create.cshtml
@@ -2,12 +2,14 @@
 @{
     ViewBag.Title = "Add New Role";
 }
-<h2>Add New User Role:</h2>
 
-@using (Html.BeginForm())
-{
-    <p> Enter User Role Name:</p>
+    <h2>Crear Nuevo Rol de Usuario:</h2>
+<div class="form-group ">
+    @using (Html.BeginForm())
+    {
+        <p> Ingrese Nuevo Rol de Usuario:</p>
 
-    @Html.EditorFor(m => m.Name)
-    <input type="submit" value="Create Role" />
-}
+        @Html.EditorFor(m => m.Name, new { htmlAttributes = new { @class = "form-control" } })
+        <input type="submit" value="Crear Rol" class="btn btn-default" />
+    }
+</div>
