+++ WebApplication8/Views/Role/Index.cshtml
@@ -2,8 +2,8 @@
 @{
     ViewBag.Title = "Add Role";
 }
-<table style=" background-color:#FFFFFF; border: dashed 3px #6D7B8D; padding: 5px;width: 99%;table-layout:fixed;" cellpadding="6" cellspacing="6">
-    <tr style="height: 30px; background-color:#336699 ; color:#FFFFFF ;border: solid 1px #659EC7;">
+<table style=" background-color:#FFFFFF; border:solid 1px; border-radius:8px; padding: 5px;width: 99%;table-layout:fixed;" cellpadding="6" cellspacing="6"; >
+    <tr style="height: 30px; background-color: #faad1e; color:#FFFFFF ;border: solid 1px  #000000;">
         <td align="center" colspan="2">
             <h2> Create User Roles</h2>
         </td>
@@ -12,21 +12,21 @@
 
     <tr>
         <td>
-            <table id="tbrole" style="width:100%; border:dotted 1px; background-color:gainsboro; padding-left:10px;">
+            <table id="tbrole" style="width:100%; border:solid 1px ; background-color:gainsboro; padding-left:10px;">
 
                 @foreach (var item in Model)
                 {
                     <tr>
-                        <td style="width:100%; border:dotted 1px;">
+                        <td style="width:100%; border:solid 1px;">
                             @item.Name
                         </td>
                     </tr>}
             </table>
         </td>
-        <td align="right" style="color:#FFFFFF;padding-right:10;">
+        <td align="right" style="color:#000000;padding-right:10;">
 
 
-            <h3>   @Html.ActionLink("Click to Create New Role", "Create", "Role") </h3>
+            <h3>   @Html.ActionLink(" Crear Nuevo Rol ", "Create", "Role", new { @class= "btn btn-default" }) </h3>
         </td>
     </tr>
 </table>
