+++ WebApplication8/Views/usuarios/Create.cshtml
@@ -21,7 +21,7 @@
            {
                new SelectListItem { Text="Seleccione Tipo de Usuario", Value="" },
                new SelectListItem { Text="Cliente", Value="Cliente" },
-                new SelectListItem { Text="Motero", Value="Motero" },
+                new SelectListItem { Text="Moto", Value="Moto" },
            }, new  { @class = "form-control" } )
